wip: tests, refactor and new cmd (del+get default)
