+++ utils/note/delete.go
@@ -1,0 +1,3 @@
+package note_test
+
+// TODO
