+++ utils/note/delete_test.go
@@ -1,0 +1,17 @@
+package note
+
+import (
+	"fmt"
+	"os"
+)
+
+func Delete(originalPath string) error {
+	o := AddMdSuffix(originalPath)
+	err := os.Remove(o)
+	if err != nil {
+		return err
+	} else {
+		fmt.Println("Deleted note: ", o)
+	}
+	return nil
+}
